services/agent_service/requirements.txt
google-generativeai>=0.8.3
chromadb
kafka-python
PyPDF2
